requirements.txt
Django==3.2.8
mysqlclient==2.0.3
PyMySQL==1.0.2
django-cors-headers==3.10.0
djangorestframework==3.12.4
python-dotenv==0.19.0
requests==2.26.0
gunicorn==20.1.0
PyJWT==2.3.0
python-json-logger==2.0.2
django-filter==21.1
django-redis==5.0.0
whitenoise==5.3.0 